mavlink: move VFR_HUD to separate stream header
